requirements.txt
# FastAPI Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Templating Engine (Server-rendered UI)
jinja2==3.1.4

# File Upload Support
python-multipart==0.0.18

# PDF Parsing
pdfplumber==0.11.4

# Google Gemini AI Integration
google-generativeai==0.8.3

# Data Validation
pydantic==2.10.3

# Database (SQLite ORM)
sqlalchemy==2.0.36

# Environment Variables Management
python-dotenv==1.0.1

# HTTP Client (for API calls)
httpx==0.28.1

# Date/Time Utilities
python-dateutil==2.9.0

